+++ vcm/downloader/subject.py
@@ -11,6 +11,7 @@ from _sha1 import sha1
 from vcm.core.networking import Connection
 from vcm.core.settings import DownloadSettings, GeneralSettings
 from vcm.core.utils import secure_filename
+from vcm.downloader.link import Chat, Kalvidres
 
 from .alias import Alias
 from .link import BaseLink, Delivery, Folder, ForumList, Resource
@@ -159,11 +160,19 @@ class Subject:
             elif "forum" in url:
                 self.logger.debug("Created Forum (subject search): %r, %s", name, url)
                 self.add_link(ForumList(name, section, url, icon_url, self))
+            elif "chat" in url:
+                self.logger.debug("Created Chat (subject search): %r, %s", name, url)
+                self.add_link(Chat(name, section, url, icon_url, self))
             elif "assign" in url:
                 self.logger.debug(
                     "Created Delivery (subject search): %r, %s", name, url
                 )
                 self.add_link(Delivery(name, section, url, icon_url, self))
+            elif "kalvidres" in url:
+                self.logger.debug(
+                    "Created Kalvidres (subject search): %r, %s", name, url
+                )
+                self.add_link(Kalvidres(name, section, url, icon_url, self))
 
         self.logger.debug("Downloading files for subject %r", self.name)