+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [unreleased]
 ### Added
 * Add `utils.Key` (returned by redesigned `utils.getch`)
+* Add support for `Chat` links.
+* Add support for `Kalvidres` links.
+
 ### Changed
 * Use links of the [virtual campus](https://campusvirtual.uva.es/my) instead of [google drive](https://drive.google.com/drive).
 * Renamed `IconType.unkown` to `IconType.unknown` (**typo**).